top: one boolean per strategy
top::BooleanWindow<Lead, Optional, Excludes, Table> and
top::BooleanSparse<Lead, Probes, Optional, Excludes, Table> replace
WindowDisjunction, SparseConjunction, SparseExclusion and WalkBlock.
The window root has two forms: Optional = OrGroup<Set> with the scored
Fill (the old disjunction, ProbedAndNot excludes resetting the score)
and Lead = fill::AllDocs with FilledAndNot excludes for pure negation,
which no longer walks every document through a sparse All lead. The
sparse root probes musts, optional and excludes, fetches score
arguments per slot and scores in batches of 32; its boost group is
probe::BoostLeaves<Leaf, N>, so probe/sparse_boost_scored.hpp and
probe::BothLeaves go. top::Posting gains an Excludes slot and absorbs
the single-term exclusion fast path.

Pruned roots (MaxScoreDisjunction, WandConjunction, PrunedPosting) are
unchanged. Five parity sets byte-identical (documents and scores),
gtests green, TOP_100 and TOP_100_COUNT harness 0.01x to 0.21x on the
21 negation-only shapes and inside the controls' band elsewhere.
