+++ libs/iresearch/include/iresearch/search/top/make.cpp
@@ -25,10 +25,8 @@
 
 #include "iresearch/index/index_reader.hpp"
 #include "iresearch/search/all_filter.hpp"
-#include "iresearch/search/boolean_query.hpp"
 #include "iresearch/search/collectors.hpp"
 #include "iresearch/search/common/all_docs_score.hpp"
-#include "iresearch/search/common/boolean_of.hpp"
 #include "iresearch/search/common/collect.hpp"
 #include "iresearch/search/common/ngram_of.hpp"
 #include "iresearch/search/common/phrase_of.hpp"
@@ -43,6 +41,7 @@
 #include "iresearch/search/term_query.hpp"
 #include "iresearch/search/top/detail/walk.hpp"
 #include "iresearch/search/top/empty.hpp"
+#include "iresearch/search/top/make_boolean.hpp"
 #include "iresearch/search/wildcard_ngram_filter.hpp"
 
 namespace irs::top {
@@ -169,64 +168,6 @@ Root::ptr Make(const AllQuery& query, const Context& ctx) {
 
 Root::ptr Make(const WildcardNGramQuery& query, const Context& ctx) {
   return MakeWildcardNGram(query, ctx);
-}
-
-Root::ptr Make(const BooleanQuery& query, const Context& ctx) {
-  const auto& segment = query.Segment();
-  const auto merge = query.MergeType();
-  const auto absorbed = query.Absorbed();
-  const std::span must = query.Terms(Occur::Must);
-  const std::span must_filters = query.Queries(Occur::Must);
-  const std::span should = query.Terms(Occur::Should);
-  const std::span should_filters = query.Queries(Occur::Should);
-  const auto min_match = query.MinShouldMatch();
-
-  const bool optional = !should.empty() || !should_filters.empty();
-  const bool only_scores = optional && min_match == 0;
-
-  if (!query.Terms(Occur::MustNot).empty() ||
-      !query.Queries(Occur::MustNot).empty()) {
-    if (auto windowed =
-          MakeWindowExclusion(query, segment, ctx, merge, absorbed)) {
-      return windowed;
-    }
-    return MakeSparseExclusion(query, segment, ctx, merge, absorbed);
-  }
-  if (must.empty() && must_filters.empty() && !only_scores) {
-    if (!optional) {
-      return MakeAll(segment, ctx, absorbed);
-    }
-    const auto uniformity = query.Uniformity(Occur::Should);
-    if (min_match == 1) {
-      if (ctx.prune && merge == ScoreMergeType::Sum && absorbed == 0) {
-        if (auto pruned = MakeMaxScoreDisjunction(
-              should, should_filters, uniformity, nullptr, nullptr, kNoBoost,
-              segment, ctx, merge)) {
-          return pruned;
-        }
-      }
-      return MakeWindowDisjunction(should, should_filters, uniformity, nullptr,
-                                   nullptr, kNoBoost, segment, ctx, merge,
-                                   absorbed);
-    }
-    if (min_match > search::kBitplaneMaxMatch) {
-      if (auto counted =
-            MakeCountThreshold(should, should_filters, uniformity, segment, ctx,
-                               merge, min_match, absorbed)) {
-        return counted;
-      }
-    }
-    return MakeBitsThreshold(should, should_filters, uniformity, segment, ctx,
-                             merge, min_match, absorbed);
-  }
-  if (ctx.prune && !optional && merge == ScoreMergeType::Sum && absorbed == 0) {
-    if (auto pruned =
-          MakeWandConjunction(must, must_filters, query.Uniformity(Occur::Must),
-                              segment, ctx, merge)) {
-      return pruned;
-    }
-  }
-  return MakeSparseConjunction(query, segment, ctx, merge, absorbed);
 }
 
 Root::ptr MakeRoot(const QueryBuilder& query, const Context& ctx) {
